+++ trellis_datamodel/models/schemas.py
@@ -93,6 +93,7 @@ class ConfigInfoResponse(BaseModel):
     exposures_default_layout: str
     bus_matrix_enabled: bool
     modeling_style: str
+    entity_prefix: List[str]
 
 
 class ManifestResponse(BaseModel):